Add correctly RecipeIngredients related models
With using Cocooned, the parameters received in
'recipe_ingredient_attributes' is now a hash table. the keys for each
Ingredient, Quantity, Unity are names 'ingredients', 'quantities',
unities' (for a reason not clear to me right now).

This brings confusion and prevent Rails to append
{#model_name}_attributes as it has been doing for recipe_ingredients.

As a workaround, creates private functions in
recipe ingredient controller to handle those weird parameters and create
the appropriate models.

At some point this part of code should be modified once we understand
how to do this in a more sustainable way.